HW-5, refactoring of a components tree
