+++ src/components/restaurant-item/restaurant-item.jsx
@@ -1,5 +1,5 @@
-import { Menu } from '../menu';
-import { Reviews } from '../reviews';
+import { Menu } from '../menu-list/menu-list';
+import { Reviews } from '../review-list/review-list';
 
 export const Restaurant = ({ id, menu, reviews }) => {
   return (
